requirements.txt
# --- Framework Web/API ---
fastapi==0.116.1
uvicorn==0.35.0

# --- Validação e Utilidades ---
pydantic==2.11.7
python-dotenv==1.1.1

# --- Integração com LLMs e IA ---
openai==1.96.1
langchain==0.3.26
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-community==0.3.27
langchain-chroma==0.2.4
chromadb==1.0.15
sentence-transformers==5.0.0

# --- Manipulação de Documentos (DOCX, PDF) ---
python-docx==1.2.0
pdfminer.six==20221105

# --- Manipulação de Dados ---
numpy==2.2.6
pandas==2.3.1

# --- Scripts de Pré-processamento ---
docling==2.41.0